+++ src/services/posts.py
@@ -1,6 +1,7 @@
 # pixels_project\src\services\posts.py
 from typing import List, Any
 from sqlalchemy.orm import Session
+
 from src.services.core import BaseServices, ModelType
 from src.database.models import Image
 from src.schemas import PostCreate, PostUpdate
